Now game parameters information is correctly loaded with scenario.
